refactor: user logs state moved to UsersTable component
