+++ src/components/SelectedWindow.tsx
@@ -4,13 +4,13 @@ import LogsTable from "./LogsTable.tsx";
 import ExampleChart from "./Chart.tsx";
 import SignIn from "./SignIn.tsx";
 
-function SelectedWindow({ userLogs }) {
+function SelectedWindow() {
   const location = useLocation();
 
   return (
     <>
       {location.pathname.includes("/notification-logs") && <LogsTable />}
-      {location.pathname === "/users" && <UsersTable logs={userLogs} />}
+      {location.pathname === "/users" && <UsersTable />}
       {location.pathname === "/analytics" && <ExampleChart />}
       {location.pathname === "/sign-in" && <SignIn />}
     </>
